Improve Features and BugFix.
Added better functionality by making the "Export Columns" feature, export with the current selected sheet name.

Improved "Concatenate Files", with the ability to select between concatenate files in a single or multi sheet Exceel File.

Fixed a bug that made the application show that we had something loaded when we closed the FileDialog without selecting anything.